Revert "Add an include folder"
This reverts commit 84882b5c682102141fd42e85601ad689c989e799.
